COM-2818: change names
