+++ src/components/DeletionConfirmationModal/index.tsx
@@ -1,39 +1,96 @@
-import React from 'react';
-import { View, Text, TouchableOpacity } from 'react-native';
+import React, { useState, useReducer, useEffect } from 'react';
+import { View, Text, TouchableOpacity, ActivityIndicator } from 'react-native';
+import { AntDesign } from '@expo/vector-icons';
 import styles from './styles';
 import NiModal from '../Modal';
+import NiInput from '../form/Input';
+import Users from '../../api/users';
 import FeatherButton from '../icons/FeatherButton';
 import { HIT_SLOP, ICON } from '../../styles/metrics';
-import { GREY } from '../../styles/colors';
+import { errorReducer, initialErrorState, RESET_ERROR, SET_ERROR } from '../../reducers/error';
+import { GREY, ORANGE } from '../../styles/colors';
+import commonStyle from '../../styles/common';
 
 interface DeletionConfirmationModalProps {
   visible: boolean,
-  name: string,
-  logout: () => void,
+  loggedUserId: string,
+  setVisible: (value) => void,
+  setConfirmationModal: () => void,
 }
 
 const DeletionConfirmationModal = ({
   visible,
-  name,
-  logout,
-}: DeletionConfirmationModalProps) => (
-  <NiModal visible={visible}>
-    <>
-      <View style={styles.header}>
-        <Text style={styles.title}>Votre compte a été supprimé</Text>
-        <FeatherButton name="x" onPress={logout} size={ICON.MD} color={GREY[600]} />
-      </View>
-      <Text style={styles.subTitle}>Nous sommes désolé de vous voir partir {name} 😔</Text>
-      <Text style={styles.body}>
-          Si vous changez d&apos;avis, vous pourrez vous recréer un compte et découvrir notre nouveau contenu
-          à tout moment.
-      </Text>
-      <View style={styles.buttonContainer}>
-        <TouchableOpacity hitSlop={HIT_SLOP} onPress={logout}>
-          <Text style={styles.button}>Ok</Text>
-        </TouchableOpacity>
-      </View>
-    </>
-  </NiModal>
-);
+  loggedUserId,
+  setVisible,
+  setConfirmationModal,
+}: DeletionConfirmationModalProps) => {
+  const [confirmationInput, setConfirmationInput] = useState<string>('');
+  const [error, dispatchError] = useReducer(errorReducer, initialErrorState);
+  const [isValidationAttempted, setIsValidationAttempted] = useState<boolean>(false);
+  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const CONFIRMATION_WORD = 'supprimer';
+
+  useEffect(() => {
+    const isConfirmationInputInvalid = !(confirmationInput.toLowerCase().trim() === CONFIRMATION_WORD);
+    if (isConfirmationInputInvalid) {
+      const payload = isValidationAttempted
+        ? `Vous devez écrire “${CONFIRMATION_WORD}” ici pour confirmer la suppression de compte`
+        : '';
+      dispatchError({ type: SET_ERROR, payload });
+    } else { dispatchError({ type: RESET_ERROR }); }
+  }, [confirmationInput, isValidationAttempted]);
+
+  const deleteAccount = async () => {
+    try {
+      setIsValidationAttempted(true);
+      if (!error.value) {
+        setIsLoading(true);
+        await Users.deleteAccount(loggedUserId);
+        setVisible(false);
+        setConfirmationModal();
+      }
+    } catch (e) {
+      dispatchError({ type: SET_ERROR, payload: 'Oops, une erreur est survenue' });
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  const onRequestClose = () => {
+    setIsValidationAttempted(false);
+    setVisible(true);
+    setConfirmationInput('');
+  };
+
+  return (
+    <NiModal visible={visible}>
+      <>
+        <View style={styles.header}>
+          <Text style={styles.title}>Êtes vous sûr de vouloir {CONFIRMATION_WORD} votre compte ?</Text>
+          <FeatherButton name="x" onPress={onRequestClose} size={ICON.MD} color={GREY[600]} />
+        </View>
+        <View style={styles.warningMessage}>
+          <AntDesign name="exclamationcircleo" color={ORANGE[700]} size={ICON.MD} style={styles.warningIcon} />
+          <Text style={styles.warningMessage}>
+            Votre historique de formation sera supprimé. Cette action est irréversible.
+          </Text>
+        </View>
+        <Text style={styles.body}>Veuillez écrire “{CONFIRMATION_WORD}” pour confirmer cette action.</Text>
+        <NiInput value={confirmationInput} onChangeText={setConfirmationInput} caption=""
+          placeholder={CONFIRMATION_WORD} type="text" borderColor={GREY[200]} validationMessage={error.message} />
+        <View style={styles.buttonContainer}>
+          <TouchableOpacity hitSlop={HIT_SLOP} onPress={onRequestClose}>
+            <Text style={styles.button}>Annuler</Text>
+          </TouchableOpacity>
+          <TouchableOpacity hitSlop={HIT_SLOP} onPress={deleteAccount}>
+            {!isLoading
+              ? <Text style={styles.button}>Supprimer mon compte</Text>
+              : <ActivityIndicator style={commonStyle.disabled} color={GREY[800]} size="small" />}
+          </TouchableOpacity>
+        </View>
+      </>
+    </NiModal>
+  );
+};
+
 export default DeletionConfirmationModal;
