+++ src/screens/profile/Profile/index.tsx
@@ -18,8 +18,8 @@ import { HIT_SLOP, ICON } from '../../../styles/metrics';
 import FeatherButton from '../../../components/icons/FeatherButton';
 import PictureModal from '../../../components/PictureModal';
 import CompanySearchModal from '../../../components/companyLinkRequest/CompanySearchModal';
-import DeletionModal from '../../../components/DeletionModal';
 import DeletionConfirmationModal from '../../../components/DeletionConfirmationModal';
+import UserAccountDeletedModal from '../../../components/UserAccountDeletedModal';
 
 interface ProfileProps extends CompositeScreenProps<
 StackScreenProps<RootBottomTabParamList>,
@@ -36,8 +36,8 @@ const Profile = ({ loggedUser, navigation }: ProfileProps) => {
   const [hasPhoto, setHasPhoto] = useState<boolean>(false);
   const [pictureModal, setPictureModal] = useState<boolean>(false);
   const [isModalOpened, setIsModalOpened] = useState<boolean>(false);
-  const [deletionModal, setDeletionModal] = useState<boolean>(false);
   const [deletionConfirmationModal, setDeletionConfirmationModal] = useState<boolean>(false);
+  const [userAccountDeletedModal, setUserAccountDeletedModal] = useState<boolean>(false);
 
   const getUserCourses = async () => {
     try {
@@ -136,7 +136,7 @@ const Profile = ({ loggedUser, navigation }: ProfileProps) => {
         <Text style={styles.legalNotice}>Conditions d’utilisation</Text>
       </TouchableOpacity>
       {!get(loggedUser, 'company') &&
-        <TouchableOpacity hitSlop={HIT_SLOP} onPress={() => setDeletionModal(true)}
+        <TouchableOpacity hitSlop={HIT_SLOP} onPress={() => setDeletionConfirmationModal(true)}
           style={styles.legalNoticeContainer}>
           <Text style={styles.legalNotice}>Supprimer mon compte</Text>
         </TouchableOpacity>}
@@ -147,10 +147,10 @@ const Profile = ({ loggedUser, navigation }: ProfileProps) => {
       <PictureModal visible={pictureModal} hasPhoto={hasPhoto} setPictureModal={setPictureModal} setSource={setSource}
         setHasPhoto={setHasPhoto} />
       <CompanySearchModal visible={isModalOpened} onRequestClose={() => setIsModalOpened(false)} />
-      <DeletionModal visible={deletionModal} setVisible={() => setDeletionModal(false)}
-        loggedUserId={get(loggedUser, '_id')}
-        setConfirmationModal={() => setDeletionConfirmationModal(true)} />
-      <DeletionConfirmationModal visible={deletionConfirmationModal} name={get(loggedUser, 'identity.firstname')}
+      <DeletionConfirmationModal visible={deletionConfirmationModal} loggedUserId={get(loggedUser, '_id')}
+        setVisible={() => setDeletionConfirmationModal(false)}
+        setConfirmationModal={() => setUserAccountDeletedModal(true)} />
+      <UserAccountDeletedModal visible={userAccountDeletedModal} name={get(loggedUser, 'identity.firstname')}
         logout={clearExpoTokenAndSignOut} />
     </ScrollView>
   );
